requirements.txt
langchain-community==0.2.11
langchain-core==0.2.28
pypdf==4.3.1
docx2txt==0.8
tiktoken==0.7.0
scikit-learn==1.5.1
faiss-cpu==1.8.0.post1
openai==1.40.0
pytest==8.3.2